♻️  Move polish notation tests to own test class
